// server/src/graphql.schema.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  DateTime: Date;
  JSON: any;
  Upload: any;
};

export type Account = {
  __typename?: 'Account';
  /** Account creation time */
  createdAt: Scalars['DateTime'];
  /** Account deletion time */
  deletedAt?: Maybe<Scalars['DateTime']>;
  /** list of documents that account is member of */
  documents: Array<AccountDocument>;
  /** Account ID */
  id: Scalars['ID'];
  /** Optional account name */
  name?: Maybe<Scalars['String']>;
  /** Account last update time */
  updatedAt: Scalars['DateTime'];
};

export type AccountDocument = {
  __typename?: 'AccountDocument';
  /** Time when account was accepted as member of document */
  acceptedAt: Scalars['DateTime'];
  /** document member creation time */
  createdAt: Scalars['DateTime'];
  /** account that created document member */
  createdBy: AccountInfo;
  /** document associated with account */
  document?: Maybe<Document>;
  /** Role of account for document */
  role: DocumentRole;
};

export type AccountInfo = {
  __typename?: 'AccountInfo';
  /** account ID */
  id: Scalars['ID'];
  /** optional account name */
  name?: Maybe<Scalars['String']>;
};

export type AccountUpdateInput = {
  /** id of the account to update */
  id: Scalars['ID'];
  /** name of the account */
  name?: InputMaybe<Scalars['String']>;
  /** update timestamp */
  timestamp: Scalars['DateTime'];
};

export type Document = {
  __typename?: 'Document';
  /** document creation time */
  createdAt: Scalars['DateTime'];
  /** account that created document */
  createdBy: AccountInfo;
  /** document deletion time */
  deletedAt?: Maybe<Scalars['DateTime']>;
  /** file associated with document */
  file?: Maybe<FileInfo>;
  /** list of highlights associated with document */
  highlights: Array<DocumentHighlight>;
  /** unique document ID */
  id: Scalars['ID'];
  /** list of document members */
  members: Array<DocumentMember>;
  /** metadata associated with document */
  meta: DocumentMeta;
  /** type of the document */
  type: DocumentType;
  /** document last update time */
  updatedAt: Scalars['DateTime'];
};


export type DocumentHighlightsArgs = {
  limit?: InputMaybe<Scalars['Int']>;
  offset?: InputMaybe<Scalars['Int']>;
};

export type DocumentChange = {
  __typename?: 'DocumentChange';
  /** whether document has been deleted */
  deleted?: Maybe<Scalars['Boolean']>;
  /** changes regarding document highlights */
  highlights?: Maybe<Array<DocumentHighlightChange>>;
  /** changes regarding document members */
  members?: Maybe<Array<DocumentMemberChange>>;
  /** whether document meta has been updated */
  meta?: Maybe<DocumentMeta>;
  /** timestamp of change */
  timestamp: Scalars['DateTime'];
};

export type DocumentHighlight = {
  __typename?: 'DocumentHighlight';
  /** content associated with highlight */
  content: Scalars['JSON'];
  /** highlight creation time */
  createdAt: Scalars['DateTime'];
  /** highlight author ID */
  createdBy: AccountInfo;
  /** highlight deletion time */
  deletedAt?: Maybe<Scalars['DateTime']>;
  /** unique highlight id */
  id: Scalars['ID'];
  /** highlight location */
  location: Scalars['JSON'];
  /** highlight last udpate time */
  updatedAt: Scalars['DateTime'];
};

export type DocumentHighlightChange = {
  __typename?: 'DocumentHighlightChange';
  /** whether highlight content has been changed */
  content: Scalars['JSON'];
  /** who has made a change */
  createdBy: AccountInfo;
  /** whether highlight has been deleted */
  deleted?: Maybe<Scalars['Boolean']>;
  /** unique highlight id */
  id: Scalars['ID'];
  /** whether highlight location has been changed */
  location: Scalars['JSON'];
};

export type DocumentMember = {
  __typename?: 'DocumentMember';
  /** Time when account was accepted as member of document */
  acceptedAt?: Maybe<Scalars['DateTime']>;
  /** member account */
  account: AccountInfo;
  /** document member creation time */
  createdAt: Scalars['DateTime'];
  /** account that created document member */
  createdBy: AccountInfo;
  /** Role of account for document */
  role: DocumentRole;
};

export type DocumentMemberChange = {
  __typename?: 'DocumentMemberChange';
  /** whether member has been accepted */
  accepted?: Maybe<Scalars['Boolean']>;
  /** account id that membership has been changed for */
  accountId: Scalars['ID'];
  /** who has made a change */
  createdBy: AccountInfo;
  /** whether member has been deleted */
  deleted?: Maybe<Scalars['Boolean']>;
  /** whether ddocument role has been changed */
  role?: Maybe<DocumentRole>;
};

export type DocumentMemberInput = {
  /** whether document member is accepted */
  accepted?: InputMaybe<Scalars['Boolean']>;
  /** id of account to update */
  accountId: Scalars['ID'];
  /** role of document member */
  role?: InputMaybe<DocumentRole>;
};

/** user provided document metadata */
export type DocumentMeta = {
  __typename?: 'DocumentMeta';
  /** document author */
  author?: Maybe<Scalars['String']>;
  /** url of document cover image */
  cover?: Maybe<Scalars['String']>;
  /** document description */
  description?: Maybe<Scalars['String']>;
  /** keywords associated with document */
  keywords?: Maybe<Array<Scalars['String']>>;
  /** number of document pages */
  pageCount?: Maybe<Scalars['Int']>;
  /** title of the document */
  title?: Maybe<Scalars['String']>;
};
// ...
export enum DocumentRole {
  Admin = 'ADMIN',
  Editor = 'EDITOR',
  Viewer = 'VIEWER'
}

export enum DocumentType {
  Pdf = 'PDF'
}

export type FileInfo = {
  __typename?: 'FileInfo';
  /** file creation time */
  createdAt: Scalars['DateTime'];
  /** account that created file */
  createdBy: AccountInfo;
  /** sha256 hash of file */
  hash: Scalars['String'];
  /** file mime type */
  mimeType: Scalars['String'];
  /** alternative file sources */
  sources?: Maybe<Array<Scalars['String']>>;
  /** file last update time */
  updatedAt: Scalars['DateTime'];
  /** primary url of file */
  url?: Maybe<Scalars['String']>;
};
// ...
export type WithIndex<TObject> = TObject & Record<string, any>;
export type ResolversObject<TObject> = WithIndex<TObject>;
// ...
export type AccountResolvers<ContextType = any, ParentType extends ResolversParentTypes['Account'] = ResolversParentTypes['Account']> = ResolversObject<{
  createdAt?: Resolver<ResolversTypes['DateTime'], ParentType, ContextType>;
  deletedAt?: Resolver<Maybe<ResolversTypes['DateTime']>, ParentType, ContextType>;
  documents?: Resolver<Array<ResolversTypes['AccountDocument']>, ParentType, ContextType>;
  id?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
  name?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  updatedAt?: Resolver<ResolversTypes['DateTime'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;

export type AccountDocumentResolvers<ContextType = any, ParentType extends ResolversParentTypes['AccountDocument'] = ResolversParentTypes['AccountDocument']> = ResolversObject<{
  acceptedAt?: Resolver<ResolversTypes['DateTime'], ParentType, ContextType>;
  createdAt?: Resolver<ResolversTypes['DateTime'], ParentType, ContextType>;
  createdBy?: Resolver<ResolversTypes['AccountInfo'], ParentType, ContextType>;
  document?: Resolver<Maybe<ResolversTypes['Document']>, ParentType, ContextType>;
  role?: Resolver<ResolversTypes['DocumentRole'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;

export type AccountInfoResolvers<ContextType = any, ParentType extends ResolversParentTypes['AccountInfo'] = ResolversParentTypes['AccountInfo']> = ResolversObject<{
  id?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
  name?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;

export interface DateTimeScalarConfig extends GraphQLScalarTypeConfig<ResolversTypes['DateTime'], any> {
  name: 'DateTime';
}

export type DocumentResolvers<ContextType = any, ParentType extends ResolversParentTypes['Document'] = ResolversParentTypes['Document']> = ResolversObject<{
  createdAt?: Resolver<ResolversTypes['DateTime'], ParentType, ContextType>;
  createdBy?: Resolver<ResolversTypes['AccountInfo'], ParentType, ContextType>;
  deletedAt?: Resolver<Maybe<ResolversTypes['DateTime']>, ParentType, ContextType>;
  file?: Resolver<Maybe<ResolversTypes['FileInfo']>, ParentType, ContextType>;
  highlights?: Resolver<Array<ResolversTypes['DocumentHighlight']>, ParentType, ContextType, Partial<DocumentHighlightsArgs>>;
  id?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
  members?: Resolver<Array<ResolversTypes['DocumentMember']>, ParentType, ContextType>;
  meta?: Resolver<ResolversTypes['DocumentMeta'], ParentType, ContextType>;
  type?: Resolver<ResolversTypes['DocumentType'], ParentType, ContextType>;
  updatedAt?: Resolver<ResolversTypes['DateTime'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;

export type DocumentChangeResolvers<ContextType = any, ParentType extends ResolversParentTypes['DocumentChange'] = ResolversParentTypes['DocumentChange']> = ResolversObject<{
  deleted?: Resolver<Maybe<ResolversTypes['Boolean']>, ParentType, ContextType>;
  highlights?: Resolver<Maybe<Array<ResolversTypes['DocumentHighlightChange']>>, ParentType, ContextType>;
  members?: Resolver<Maybe<Array<ResolversTypes['DocumentMemberChange']>>, ParentType, ContextType>;
  meta?: Resolver<Maybe<ResolversTypes['DocumentMeta']>, ParentType, ContextType>;
  timestamp?: Resolver<ResolversTypes['DateTime'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;

export type DocumentHighlightResolvers<ContextType = any, ParentType extends ResolversParentTypes['DocumentHighlight'] = ResolversParentTypes['DocumentHighlight']> = ResolversObject<{
  content?: Resolver<ResolversTypes['JSON'], ParentType, ContextType>;
  createdAt?: Resolver<ResolversTypes['DateTime'], ParentType, ContextType>;
  createdBy?: Resolver<ResolversTypes['AccountInfo'], ParentType, ContextType>;
  deletedAt?: Resolver<Maybe<ResolversTypes['DateTime']>, ParentType, ContextType>;
  id?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
  location?: Resolver<ResolversTypes['JSON'], ParentType, ContextType>;
  updatedAt?: Resolver<ResolversTypes['DateTime'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;

export type DocumentHighlightChangeResolvers<ContextType = any, ParentType extends ResolversParentTypes['DocumentHighlightChange'] = ResolversParentTypes['DocumentHighlightChange']> = ResolversObject<{
  content?: Resolver<ResolversTypes['JSON'], ParentType, ContextType>;
  createdBy?: Resolver<ResolversTypes['AccountInfo'], ParentType, ContextType>;
  deleted?: Resolver<Maybe<ResolversTypes['Boolean']>, ParentType, ContextType>;
  id?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
  location?: Resolver<ResolversTypes['JSON'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;

export type DocumentMemberResolvers<ContextType = any, ParentType extends ResolversParentTypes['DocumentMember'] = ResolversParentTypes['DocumentMember']> = ResolversObject<{
  acceptedAt?: Resolver<Maybe<ResolversTypes['DateTime']>, ParentType, ContextType>;
  account?: Resolver<ResolversTypes['AccountInfo'], ParentType, ContextType>;
  createdAt?: Resolver<ResolversTypes['DateTime'], ParentType, ContextType>;
  createdBy?: Resolver<ResolversTypes['AccountInfo'], ParentType, ContextType>;
  role?: Resolver<ResolversTypes['DocumentRole'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;
// ...
export type Resolvers<ContextType = any> = ResolversObject<{
  Account?: AccountResolvers<ContextType>;
  AccountDocument?: AccountDocumentResolvers<ContextType>;
  AccountInfo?: AccountInfoResolvers<ContextType>;
  DateTime?: GraphQLScalarType;
  Document?: DocumentResolvers<ContextType>;
  DocumentChange?: DocumentChangeResolvers<ContextType>;
  DocumentHighlight?: DocumentHighlightResolvers<ContextType>;
  DocumentHighlightChange?: DocumentHighlightChangeResolvers<ContextType>;
  DocumentMember?: DocumentMemberResolvers<ContextType>;
  DocumentMemberChange?: DocumentMemberChangeResolvers<ContextType>;
  DocumentMeta?: DocumentMetaResolvers<ContextType>;
  DocumentReadingInfo?: DocumentReadingInfoResolvers<ContextType>;
  FileInfo?: FileInfoResolvers<ContextType>;
  JSON?: GraphQLScalarType;
  Mutation?: MutationResolvers<ContextType>;
  Query?: QueryResolvers<ContextType>;
  Subscription?: SubscriptionResolvers<ContextType>;
  Upload?: GraphQLScalarType;
}>;